+++ CHANGELOG.rst
@@ -10,7 +10,9 @@ Version 1.5.0
 * Move framers from transaction.py to respective modules
 * Fix modbus payload builder and decoder
 * Async servers can now have an option to defer `reactor.run()` when using `Start<Tcp/Serial/Udo>Server(...,defer_reactor_run=True)`
-* Fix UDP client issue while handling MEI messages (ReadDeviceInformationRequest) 
+* Fix UDP client issue while handling MEI messages (ReadDeviceInformationRequest)
+* Add expected response lengths for WriteMultipleCoilRequest and WriteMultipleRegisterRequest
+* Fix struct errors while decoding stray response
 * Fix Misc examples
 
 Version 1.4.0